Health saves on switch, so many bug Fixes
BIG BOY BUG: Somehow when the player switches EnemyAI knows without being told. This allows the EnemyAI to "predict" the switch and hit player for super effective damage

Implemented health saving during switch, but since all data is saved in Scriptable Objects you cannot have more than one of the same kind of Goblinmon on the field at a time, add system to create "clone" of the SO so that they can be used seperately?