Party Saving between Scenes
Left two TODO's in Switching Manager, one for SavePlayerData() and the other in PopulateUnits(). These need to be adjusted so that they use PartyStorage instead of their internal goblinmon list from the previous system. I would do it tonight but I have a sleep schedule to keep to so I will fix it in the morning.
Eat shit future me!!!!!!!
-Past You