Fusion completed and Bug Tested
Fusion functions properly. Added rearranging party members in overworld. Adjusted UI (again, if I have to do it again I will be on suicide watch [this is a joke not a cry for help]) for 1920x1080. Fusion Calculator still needs to be populated with all fusions but that's a tomorrow me problem not a today me problem. Next week will probably be all Unit creation (moves, levels, hp, etc.).