New Tileset & ANIMATION FIX!!!
After literal HOURS of diagnosing this stupid fucking but the players walking animation now plays correctly. The fix is, frankly, really fucking stupid, but I'm tired of thinking about it. Added ANOTHER new tileset that I hope to God will be the final one (9 hours on Aesprite as of writing this). Fuck my chungus life dude.